# documentation/backend/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Alforis CRM API",
    "version": "1.0.0",
    "description": "Documentation de référence de l'API CRM Alforis. Les endpoints décrits couvrent les fonctionnalités principales livrées dans le plan d'amélioration : authentification, organisations, personnes, tâches, mandats, recherche globale, exports et webhooks.",
    "contact": {
      "name": "Équipe CRM Alforis",
      "email": "tech@alforis-finance.com"
    }
  },
  "servers": [
    {
      "url": "https://crm.alforis-finance.com",
      "description": "Production"
    },
    {
      "url": "http://localhost:8000",
      "description": "Développement local"
    }
  ],
...
  "paths": {
...
    "/api/v1/tasks": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "Lister les tâches",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1000
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "todo",
                "doing",
                "done",
                "snoozed"
              ]
            }
          },
          {
            "name": "priority",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "critique",
                "haute",
                "moyenne",
                "basse",
                "non_prioritaire"
              ]
            }
          },
          {
            "name": "category",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "relance",
                "rdv",
                "email",
                "due_diligence",
                "admin",
                "pitch",
                "negociation",
                "autre"
              ]
            }
          },
          {
            "name": "view",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "today",
                "overdue",
                "next7",
                "all"
              ]
            }
          },
          {
            "name": "investor_id",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "organisation_id",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "person_id",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Page de tâches",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedTasks"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Créer une tâche",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tâche créée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/v1/tasks/stats": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "Statistiques des tâches",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Stats agrégées",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskStats"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/v1/tasks/{task_id}": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "Détail d'une tâche",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tâche détaillée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "put": {
        "tags": [
          "tasks"
        ],
        "summary": "Mettre à jour une tâche",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tâche mise à jour",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "tags": [
          "tasks"
        ],
        "summary": "Supprimer une tâche",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Supprimé"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/v1/tasks/{task_id}/snooze": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Snoozer une tâche",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskSnoozeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tâche snoozée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/v1/tasks/{task_id}/quick-action": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Action rapide",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskQuickActionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tâche mise à jour",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Task": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "due_date": {
            "type": "string",
            "format": "date"
          },
          "snoozed_until": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "completed_at": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "priority": {
            "type": "string",
            "enum": [
              "critique",
              "haute",
              "moyenne",
              "basse",
              "non_prioritaire"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "todo",
              "doing",
              "done",
              "snoozed"
            ]
          },
          "category": {
            "type": "string",
            "enum": [
              "relance",
              "rdv",
              "email",
              "due_diligence",
              "admin",
              "pitch",
              "negociation",
              "autre"
            ]
          },
          "investor_id": {
            "type": "integer",
            "nullable": true
          },
          "fournisseur_id": {
            "type": "integer",
            "nullable": true
          },
          "organisation_id": {
            "type": "integer",
            "nullable": true
          },
          "person_id": {
            "type": "integer",
            "nullable": true
          },
          "is_auto_created": {
            "type": "boolean"
          },
          "auto_creation_rule": {
            "type": "string",
            "nullable": true
          },
          "is_overdue": {
            "type": "boolean"
          },
          "is_today": {
            "type": "boolean"
          },
          "is_next_7_days": {
            "type": "boolean"
          },
          "days_until_due": {
            "type": "integer"
          },
          "investor_name": {
            "type": "string",
            "nullable": true
          },
          "fournisseur_name": {
            "type": "string",
            "nullable": true
          },
          "person_name": {
            "type": "string",
            "nullable": true
          },
          "linked_entity_display": {
            "type": "string",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "required": [
          "id",
          "title",
          "due_date",
          "priority",
          "status",
          "category",
          "is_auto_created",
          "is_overdue",
          "is_today",
          "is_next_7_days",
          "days_until_due"
        ]
      },
      "TaskCreate": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "due_date": {
            "type": "string",
            "format": "date"
          },
          "priority": {
            "type": "string",
            "enum": [
              "critique",
              "haute",
              "moyenne",
              "basse",
              "non_prioritaire"
            ],
            "default": "moyenne"
          },
          "status": {
            "type": "string",
            "enum": [
              "todo",
              "doing",
              "done",
              "snoozed"
            ],
            "default": "todo"
          },
          "category": {
            "type": "string",
            "enum": [
              "relance",
              "rdv",
              "email",
              "due_diligence",
              "admin",
              "pitch",
              "negociation",
              "autre"
            ],
            "default": "autre"
          },
          "investor_id": {
            "type": "integer",
            "nullable": true
          },
          "fournisseur_id": {
            "type": "integer",
            "nullable": true
          },
          "organisation_id": {
            "type": "integer",
            "nullable": true
          },
          "person_id": {
            "type": "integer",
            "nullable": true
          },
          "is_auto_created": {
            "type": "boolean",
            "default": false
          },
          "auto_creation_rule": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "title",
          "due_date"
        ]
      },
      "TaskUpdate": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "due_date": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "priority": {
            "type": "string",
            "enum": [
              "critique",
              "haute",
              "moyenne",
              "basse",
              "non_prioritaire"
            ],
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "todo",
              "doing",
              "done",
              "snoozed"
            ],
            "nullable": true
          },
          "category": {
            "type": "string",
            "enum": [
              "relance",
              "rdv",
              "email",
              "due_diligence",
              "admin",
              "pitch",
              "negociation",
              "autre"
            ],
            "nullable": true
          },
          "investor_id": {
            "type": "integer",
            "nullable": true
          },
          "fournisseur_id": {
            "type": "integer",
            "nullable": true
          },
          "organisation_id": {
            "type": "integer",
            "nullable": true
          },
          "person_id": {
            "type": "integer",
            "nullable": true
          }
        }
      },
      "TaskSnoozeRequest": {
        "type": "object",
        "properties": {
          "days": {
            "type": "integer",
            "minimum": 1,
            "maximum": 365
          }
        },
        "required": [
          "days"
        ]
      },
      "TaskQuickActionRequest": {
        "type": "object",
        "properties": {
          "action": {
            "type": "string",
            "enum": [
              "snooze_1d",
              "snooze_1w",
              "mark_done",
              "next_day"
            ]
          }
        },
        "required": [
          "action"
        ]
      },
      "TaskStats": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer"
          },
          "overdue": {
            "type": "integer"
          },
          "today": {
            "type": "integer"
          },
          "next_7_days": {
            "type": "integer"
          },
          "by_status": {
            "type": "object",
            "additionalProperties": {
              "type": "integer"
            }
          },
          "by_priority": {
            "type": "object",
            "additionalProperties": {
              "type": "integer"
            }
          }
        },
        "required": [
          "total",
          "overdue",
          "today",
          "next_7_days",
          "by_status",
          "by_priority"
        ]
      },
      "PaginatedTasks": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer"
          },
          "skip": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Task"
            }
          }
        },
        "required": [
          "total",
          "skip",
          "limit",
          "items"
        ]
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Auth requise",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "detail": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "ValidationError": {
        "description": "Erreur de validation",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "detail": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "loc": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      },
                      "msg": {
                        "type": "string"
                      },
                      "type": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "NotFound": {
        "description": "Ressource introuvable",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "detail": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  }
}